Resolved PR requests about commenting and refactor winner displayer into different ts file JC-622
