+++ apps/react/src/store/genre/slice.ts
@@ -3,7 +3,7 @@ import { createSlice } from '@reduxjs/toolkit';
 import { fetchGenres } from './dispatchers';
 import { initialState } from './state';
 
-/** JSDoc . */
+/** JSDoc. */
 export const genresSlice = createSlice({
 	name: 'genres',
 	initialState,
