+++ apps/vanilla/src/pages/blackjack/Classes/Debugger.ts
@@ -1,14 +1,15 @@
-import { Subscriber } from '../Types/subscriber';
+import { Subscriber } from '../types/subscriber';
 
-import { PlayerTurnResult } from '../Types/player-turn-result';
+import { PlayerTurnResult } from '../types/player-turn-result';
 
 import { Inspector } from './inspector';
 import { ResultData } from './result-displayer';
 
-/** ToDo .*/
+/** Debugger.*/
 export class Debugger extends Inspector implements Subscriber<PlayerTurnResult> {
 
-	/** Debugger update the dice result base on the DiceGenerator.
+	/**
+	 * Debugger update the dice result base on the DiceGenerator.
 	 * @param playerDiceResult PlayerDuceResult is the dice result of current player.
 	 */
 	public update(playerDiceResult: PlayerTurnResult): void {
