+++ apps/vanilla/src/pages/blackjack/guards/isHTMLButtonElement.ts
@@ -1,4 +1,5 @@
-/** Type Guard to return if element is HTML Element.
+/**
+ * Type Guard to return if element is HTML Element.
  * @param element Input Element.
  */
 export function isHTMLButtonElement(element: unknown): element is HTMLButtonElement {
