+++ apps/vanilla/src/pages/blackjack/guards/isHTMLElement.ts
@@ -1,4 +1,5 @@
-/** Type Guard to return if element is HTML Element.
+/**
+ * Type Guard to return if element is HTML Element.
  * @param element Input Element.
  */
 export function isHTMLElement(element: ChildNode | null): element is HTMLElement {
