+++ apps/vanilla/src/pages/blackjack/index.css
@@ -12,8 +12,6 @@ body {
 }
 
 .container {
-	--container-box-shadow: 0 0 10px rgba(0 0 0/ 10%);
-
 	padding: var(--spacing-default);
 	box-shadow: var(--container-box-shadow);
 }
@@ -78,16 +76,16 @@ body {
 	padding: 10px 20px;
 	border: none;
 	border-radius: 5px;
-	background-color: #ff6347;
-	color: white;
+	background-color: var(--primary-color);
+	color: var(--surface-color);
 	cursor: pointer;
 }
 
 .button-roll:hover {
-	background-color: #ff4500;
+	background-color: var(--primary-color-overlay);
 }
 
 .button-roll:disabled {
-	background-color: #ddd;
-	color: black;
+	background-color: var(--surface-color);
+	color: var(--on-surface-color);
 }
